api-collection/requirements.txt
numpy==1.18.2
psycopg2-binary==2.8.4
python-dotenv==0.12.0
python-rapidjson==0.9.1